requirements.txt
asgiref==3.3.1
boto3==1.16.59
botocore==1.19.59
dj-database-url==0.5.0
Django==3.1.5
django-cors-headers==3.6.0
django-environ==0.4.5
django-storages==1.11.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
gunicorn==20.0.4
jmespath==0.10.0
psycopg2-binary==2.8.6
PyJWT==2.0.0
python-dateutil==2.8.1
pytz==2020.5
s3transfer==0.3.4
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
